+++ src/modules/exchange/dto/exchange.dto.ts
@@ -1,4 +1,4 @@
-export interface RateResponce {
+export interface RateResponse {
   conversion_rates: {
     USD: number;
     UAH?: number;
