+++ src/modules/exchange/exchangeProviders/exchange-rate.service.ts
@@ -1,11 +1,11 @@
-import { BaseExchangeService, ExchangeRateData } from '../echange.abstract';
-import { RateResponce } from '../dto/exchange.dto';
+import { BaseExchangeService, ExchangeProvider, ExchangeRateData } from '../echange.base';
+import { RateResponse } from '../dto/exchange.dto';
 
 export class ExchangeRateService extends BaseExchangeService {
   private exchangeRateAPIKey = process.env.EXCHANGE_API_KEY;
 
   constructor(nextExchangeService?: BaseExchangeService, requestURL?: string) {
-    super(nextExchangeService, requestURL);
+    super(ExchangeProvider.ExchangeRate, nextExchangeService, requestURL);
 
     this.requestURL = `https://v6.exchangerate-api.com/v6/${this.exchangeRateAPIKey}/latest/UAH`;
   }
@@ -14,7 +14,7 @@ export class ExchangeRateService extends BaseExchangeService {
    * Function to get the current USD to UAH rate
    * @returns
    */
-  public async getUsdUahRate(): Promise<RateResponce> {
+  public async getUsdUahRate(): Promise<RateResponse> {
     const data = (await this.getCurrencyData()) as ExchangeRateData;
 
     return {
