+++ src/modules/exchange/exchangeProviders/privat-bank.service.ts
@@ -1,9 +1,9 @@
-import { BaseExchangeService, PrivatBankData } from '../echange.abstract';
-import { RateResponce } from '../dto/exchange.dto';
+import { BaseExchangeService, ExchangeProvider, PrivatBankData } from '../echange.base';
+import { RateResponse } from '../dto/exchange.dto';
 
 export class PrivatBankService extends BaseExchangeService {
   constructor(nextExchangeService?: BaseExchangeService, requestURL?: string) {
-    super(nextExchangeService, requestURL);
+    super(ExchangeProvider.PrivatBank, nextExchangeService, requestURL);
 
     this.requestURL = 'https://api.privatbank.ua/p24api/pubinfo?exchange&coursid=5';
   }
@@ -12,7 +12,7 @@ export class PrivatBankService extends BaseExchangeService {
    * Function to get the current USD to UAH rate
    * @returns
    */
-  public async getUsdUahRate(): Promise<RateResponce> {
+  public async getUsdUahRate(): Promise<RateResponse> {
     const data = (await this.getCurrencyData()) as PrivatBankData[];
 
     return {
